requirements.txt
blinker==1.9.0
click==8.3.1
flask==3.1.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
werkzeug==3.1.4
yt-dlp
moviepy
Pillow
python-dotenv
